requirements.txt
streamlit
torch
transformers
opencv-python-headless
numpy
pandas
pillow
pytesseract
pdf2image
timm